### 2024-07-02 #6945 @ymzayek
Hi @kgd-al, thank you for submitting to JOSS. I am the handling editor for this submission.

Do you have any suggestion for reviewers?
- You can use the [potential reviewers list](https://reviewers.joss.theoj.org/lookup) provided at the top of this issue.
- You can also have a look at the authors in similar papers in JOSS (see https://github.com/openjournals/joss-reviews/issues/6945#issuecomment-2203404769) to look for potential reviewers.

If you have suggestions, please provide a list of potential reviewers here but do not use the @ character to avoid receiving unwanted notifications.

And finally, can you check the missing DOI listed here: https://github.com/openjournals/joss-reviews/issues/6945#issuecomment-2203405335 ?

Thank you in advance.
