### 2024-07-03 #6945 @kgd-al
Hi @ymzayek and thanks for handling to the edition of this submission.

I do not have any specific suggestion for the reviewers but based on the articles editorialbot listed:
- lutzhamel (Machine Learning and review on a gym environment)
- KanishAnand (also because of Machine Learning)
- gonsie (review of a gym environment although not from AI, I think?)
- abhiramm7 (Reinforcement Learning)
- seba-1511 (Also RL)

The missing doi is coming from  a book for which there seems to only be an ISBN. How should I proceed?
