### 2024-07-03 #6945 @ymzayek
@kgd-al ok for the missing DOI. Thanks for the list!
