### 2024-07-03 #6945 @ymzayek
Hello @lutzhamel, @KanishAnand and @gonsie ,

Would you be willing to assist in reviewing this submission for JOSS (Journal of Open Source Software)?

AMaze: a benchmark generator for sighted maze-navigating agents

You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/6945).

Thank you in advance!
