+++ src/amaze/visu/widgets/maze.py
@@ -7,8 +7,8 @@ from PyQt5.QtCore import Qt, QPointF, QRectF, QSize
 from PyQt5.QtGui import QPainter, QColor, QPainterPath, QImage
 from PyQt5.QtWidgets import QLabel
 
-from amaze.misc.resources import SignType, qimage_to_numpy, qt_images
-from amaze.misc.utils import has_qt_application, qt_application
+from ...misc.resources import SignType, qimage_to_numpy, qt_images
+from ...misc.utils import has_qt_application, qt_application
 from ._trajectory_plotter import plot_trajectory
 from ..maze import MazePainter, Color, logger
 from ...simu.maze import Maze
@@ -30,8 +30,8 @@ class QtPainter(MazePainter):
             c2 = Qt.green
             c3 = Qt.blue
         self.colors = {
-            Color.FOREGROUND: Qt.white if dark else Qt.black,
-            Color.BACKGROUND: Qt.black if dark else Qt.white,
+            Color.FOREGROUND: config["foreground"],
+            Color.BACKGROUND: config["background"],
             Color.START: c1,
             Color.FINISH: c2,
             Color.PATH: c3,
@@ -55,12 +55,12 @@ class QtPainter(MazePainter):
         alpha: float = 1,
     ):
         r = QRectF(x, y, w, h)
-        if fill:
+        if fill:  # pragma: no branch
             fc = QColor(self.colors[fill])
             if alpha < 1:
                 fc.setAlphaF(alpha)
             self.painter.fillRect(r, fc)
-        if draw:
+        if draw:  # pragma: no branch
             self.painter.setPen(self.colors[draw])
             self.painter.drawRect(r)
 
@@ -99,11 +99,9 @@ class MazeWidget(QLabel):
 
     def __init__(
         self,
-        maze: Optional[Union[Maze, Maze.BuildData, str]] = None,
-        robot: Optional[Robot] = None,
+        maze: Union[Maze, Maze.BuildData, str],
+        robot: Robot,
         config: Optional[Dict[str, Any]] = None,
-        resolution: Optional[int] = 15,
-        input_type: Optional[InputType] = InputType.DISCRETE,
         is_reset=False,
     ):
 
@@ -111,31 +109,17 @@ class MazeWidget(QLabel):
         if not is_reset:
             super().__init__()
 
-        if maze is None:
-            maze = ""
-
         self._maze = self.__to_maze(maze)
 
+        assert isinstance(robot, Robot)
         self._robot = robot
 
-        if robot is not None:
-            self._inputs = robot.data.inputs
-        else:
-            self._inputs = input_type
-
         self._clues, self._lures, self._traps = None, None, None
-        if self._inputs is InputType.CONTINUOUS:
-            if robot is not None:
-                self._vision = robot.data.vision
-            else:
-                self._vision = resolution
-        else:
-            self._vision = None
 
-        r = self._vision
-        if self._inputs is InputType.DISCRETE or r is None:
+        r = self.vision
+        if self.inputs is InputType.DISCRETE or r is None:
             m = self._maze
-            r = min(self.width(), self.height()) // min(m.width, m.height)
+            r = max(3, min(self.width(), self.height()) // min(m.width, m.height))
         else:
             r -= 2
 
@@ -166,8 +150,6 @@ class MazeWidget(QLabel):
             maze=maze,
             robot=self._robot,
             config=self._config,
-            resolution=self._vision,
-            input_type=self._inputs,
             is_reset=True,
         )
         self.update()
@@ -181,6 +163,12 @@ class MazeWidget(QLabel):
         elif not isinstance(maze, Maze):
             raise ValueError("Invalid maze argument." " Expecting Maze, Maze.BuildData or str")
[…]
