+++ pyproject.toml
@@ -75,7 +75,7 @@ docs = [
 
 [tool.pytest.ini_options]
 minversion = "6.0"
-addopts = ["-a", "--showlocals", "--strict-markers", "--strict-config"]
+addopts = ["-ra", "--showlocals", "--strict-markers", "--strict-config"]
 xfail_strict = true
 filterwarnings = ["error"]
 testpaths = ["tests"]
