### 2024-07-15 #6945 @ymzayek
Hello @abhiramm7 and @seba-1511,

Would you be willing to assist in reviewing this submission for JOSS (Journal of Open Source Software)?

AMaze: a benchmark generator for sighted maze-navigating agents

You can find more information at the top of this Github issue (https://github.com/openjournals/joss-reviews/issues/6945).

Thank you in advance!
