+++ docs/latex/benchmarking/generate_all.sh
@@ -1,5 +1,8 @@
 #!/bin/bash
 
+set -euo pipefail
+shopt -s inherit_errexit
+
 folder=$(dirname $0)
 tmp=$folder/___cache___/
 downloads=$tmp/downloads
@@ -7,12 +10,27 @@ log=$tmp/log
 
 workers=(
     "amaze amaze-benchmarker"
-    "gymnasium gymnasium"
+    "gymnasium gymnasium[classic-control,box2d,mujoco,atari,accept-rom-license]"
 )
+
+if [ $# -ge 1 ]
+then
+    echo "Detailed!"
+    workers+=(
+        "procgen procgen"
+    )
+fi
 
 cols=$(($(tput cols) - 2))
 short_output(){
-    tee -a $log - 2>&1 | stdbuf -o0 awk -vc=$cols '{printf "%s\r", substr($0, 0, c);}'
+    tee -a $log - 2>&1 | stdbuf -o0 awk -vc=$cols '
+        {
+            printf "%s", substr($0, 0, c);
+            for (i=length; i<c; i++) printf " ";
+            printf "\r";
+        }'
+    printf " %.0s" $(seq $cols)
+    printf "\r"
 }
 
 date > $log
@@ -22,15 +40,27 @@ do
     line | tee -a $log
     read worker package <<< $data
     echo "worker: $worker"
-    echo "package: $package"
     venv=$tmp/venvs/__$worker
-    python -m venv $venv
-    source $venv/bin/activate
 
-    python -m pip download -d $downloads $package | short_output
-    python -m pip install --upgrade --find-links=$downloads $package | short_output
+    if [ ! -d $venv ]
+    then
+        python -m venv $venv
+        source $venv/bin/activate
 
-    pip list
+        packages=$(tr ';' ' ' <<< $package)
+        echo "package: $package"
+
+        python -m pip download -d $downloads $packages | short_output
+        python -m pip install --upgrade --find-links=$downloads $packages pandas rich \
+        | short_output
+
+        pip list
+
+    else
+        source $venv/bin/activate
+    fi
+
+    $folder/generate_one.py $worker $@
 
 done
 line
