+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
 name = "amaze-benchmarker"
-version = "1.0rc1"
+version = "1.0rc1-post1"
 authors = [{ name="Kevin Godin-Dubois", email="k.j.m.godin-dubois@vu.nl"}]
 description = "Benchmark generator for sighted maze-navigating agents"
 readme = "README.md"
