+++ docs/latex/benchmarking/common.py
@@ -1,6 +1,7 @@
 import os
 import math
 import time
+import traceback
 
 from rich.progress import Progress as RichProgress
 
@@ -12,40 +13,63 @@ def line(c='='):
     print(c*os.get_terminal_size()[0])
 
 
-def evaluate(df, library, family, name, fn, **kwargs):
-    if name not in df.index or math.isnan(df.loc[name]["Time"]):
-        try:
-            start = time.time_ns()
-            for _ in range(REPLICATES):
-                fn(**kwargs)
-            t = (time.time_ns() - start) / (REPLICATES * 10**9)
-
-            if not math.isnan(t):
-                df.loc[name] = [library, family, t]
-
-        except Exception as e:
-            print(f"[ERROR] Processing {name}:\n{e}")
-            return False
-
-
 class Progress(RichProgress):
-    def __init__(self, base_name):
+    def __init__(self, df, library):
         super().__init__(expand=True)
-        self.base_name = base_name
-        self.description = f"[green][{self.base_name:^9}] Timing"
-        self.total = None
+        self.df, self.library = df, library
+        self.description = f"[green][{self.library:^9}] Timing"
+        self.environments = None
+        self.steps = None
+        self.errors = None
+        self.task = None
 
     def add_task(self, count):
-        self.total = count
-        return super().add_task(description=self.description, total=count)
+        self.environments = count
+        self.steps = count * REPLICATES
+        self.errors = 0
+        self.task = super().add_task(description=self.description, total=self.steps)
+        return self.task
 
-    def update(self, task, subtask):
-        if subtask is not None:
-            super().update(task, advance=1, description=f"{self.description}: {subtask}",
-                           refresh=True)
+    def close(self):
+        if self.errors > 0:
+            total = self.environments - self.errors
+            color = "red"
         else:
-            super().update(
-                task,
-                description=f"[blue][{self.base_name:^9}] Evaluated {self.total} environments",
-                advance=0, refresh=True)
+            total = self.environments
+            color = "blue"
+
+        super().update(
+            self.task,
+            description=f"[{color}][{self.library:^9}] Evaluated {total} environments",
+            advance=0, refresh=True)
+
+        if self.errors > 0:
+            super().console.print(f"[{color}] {self.errors} threw errors")
+
+    def evaluate(self, family, name, fn, label=None, **kwargs):
+        previously_completed = super().tasks[self.task].completed
+        if name not in self.df.index or math.isnan(self.df.loc[name]["Time"]):
+            try:
+                start = time.time_ns()
+                for i in range(REPLICATES):
+                    fn(**kwargs)
+
+                    super().update(self.task, advance=1,
+                                   description=f"{self.description}:"
+                                               f" {name} [{i}/{REPLICATES}]",
+                                   refresh=True)
+
+                t = (time.time_ns() - start) / (REPLICATES * 10**9)
+
+                if not math.isnan(t):
+                    self.df.loc[name] = [self.library, family, label, t]
+
+            except Exception as e:
+                print(traceback.format_exc())
+                print(f"[ERROR] Processing {name}:\n{e}")
+                self.errors += 1
+
+        super().update(self.task, completed=previously_completed+REPLICATES,
+                       description=f"{self.description}: {name}",
+                       refresh=True)
 
