+++ docs/latex/benchmarking/workers/amaze_worker.py
@@ -1,9 +1,12 @@
 import math
 import itertools
+import pprint
+
 from amaze import Maze, Robot, Simulation
 from amaze.simu.controllers import RandomController
 
-from common import evaluate, Progress, STEPS
+from common import Progress, STEPS
+
 
 def _evaluate(maze, robot):
     #help(Simulation)
@@ -19,24 +22,24 @@ def _evaluate(maze, robot):
             simulation.reset()
 
 
-def process(df, detailed):
-    amaze_fields = ["DHC"]
-    if detailed:
-        amaze_fields.extend([[11, 15, 21, 37], [5, 10, 20, 50], [0, .1, .5, 1]])
-    else:
-        amaze_fields.extend([[11, 15, 21], [5, 10, 20], [0, .5, 1]])
+def process(df):
+    visions = [11, 15, 21, 37]
+    sizes = [5, 10, 20, 50]
+    p_signs = [0, .1, .5, 1]
+    amaze_fields = [
+        *itertools.product("D", visions[:1], sizes, p_signs, p_signs),
+        *itertools.product("HC", visions, sizes, p_signs, p_signs[:1])
+    ]
 
-    with Progress("AMaze") as progress:
-        task = progress.add_task(math.prod(len(f) for f in amaze_fields))
+    with Progress(df, "AMaze") as progress:
+        progress.add_task(len(amaze_fields))
 
-        for robot, vision, size, p_lure in itertools.product(*amaze_fields):
-            namespace = "AMaze"
-            if detailed:
-                namespace += f"-{robot}"
-            mbd = Maze.BuildData.from_string(f"M4_{size}x{size}_C1_l{p_lure}_L.25")
+        for robot, vision, size, p_lure, p_trap in amaze_fields:
+            mbd = Maze.BuildData.from_string(f"M4_{size}x{size}_C1_l{p_lure}_L.25_t{p_trap}_T.5")
             rbd = Robot.BuildData.from_string(f"{robot}{vision}")
             name = f"{mbd.to_string()}-{rbd.to_string()}"
-            evaluate(df, "AMaze", namespace, name, _evaluate, maze=mbd, robot=rbd)
-            progress.update(task, name)
+            progress.evaluate("AMaze", name, _evaluate, label=robot, maze=mbd, robot=rbd)
 
-        progress.update(task, None)
+        progress.close()
+
+    return progress.errors
