### 2024-07-21 4ed7e7e
Added flag for empty intersections in input evaluation
