### 2024-09-02 #6945 @kgd-al
@ymzayek Well according to the guidelines at [the top](https://github.com/openjournals/joss-reviews/issues/6945#issuecomment-2203404769), we should consider the authors from the 5 most similar papers.
Meaning:
- rusu24edward
- stheid
- soraxas
- auroreloisy
- inpefess
