### 2024-09-10 #7208 @editorialbot
[REVIEW]: AMaze: a benchmark generator for sighted maze-navigating agents
**Submitting author:** @kgd-al (<a href="http://orcid.org/0009-0002-6033-3555">Kevin Godin - Dubois</a>)
**Repository:** https://github.com/kgd-al/amaze
**Branch with paper.md** (empty if default branch): joss
**Version:** v1.0
