### 2024-09-10 #7208 @stheid
I find the appreciation of related work a little shallow. I do not know specific publications but I doubt that there is no other benchmark that can generate simple 2d mazes.

There is only a small paragraph about contributing. Not sure if this is sufficient.  https://amaze.readthedocs.io/en/latest/#contributing
