+++ joss/paper.bib
@@ -12,6 +12,25 @@ volume = {SMC-13},
 year = {1983}
 }
 
+@article{Beattie2016,
+title = {{DeepMind} {Lab}},
+url = {https://arxiv.org/abs/1612.03801v2},
+urldate = {2024-01-26},
+author = {Beattie, Charles and Leibo, Joel Z. and Teplyashin, Denis and Ward, Tom and Wainwright, Marcus and Küttler, Heinrich and Lefrancq, Andrew and Green, Simon and Valdés, Víctor and Sadik, Amir and Schrittwieser, Julian and Anderson, Keith and York, Sarah and Cant, Max and Cain, Adam and Bolton, Adrian and Gaffney, Stephen and King, Helen and Hassabis, Demis and Legg, Shane and Petersen, Stig},
+month = dec,
+year = {2016},
+arXiv = {1612.03801},
+}
+
+@article{Beattie2020,
+title = {{DeepMind} {Lab2D}},
+url = {https://github.com/deepmind/lab2d},
+urldate = {2024-01-26},
+author = {Beattie, Charles and Köppe, Thomas and Duéñez-Guzmán, Edgar A and Leibo, Joel Z},
+year = {2020},
+arXiv = {2011.07027},
+}
+
 @article{Bellemare2013,
 author = {Bellemare, Marc G. and Naddaf, Yavar and Veness, Joel and Bowling, Michael},
 doi = {10.1613/jair.3912},
@@ -24,6 +43,42 @@ title = {{The Arcade Learning Environment: An Evaluation Platform for General Ag
 url = {https://jair.org/index.php/jair/article/view/10819},
 volume = {47},
 year = {2013}
+}
+
+@inproceedings{Harries2019,
+title = {{MazeExplorer}: {A} {Customisable} {3D} {Benchmark} for {Assessing} {Generalisation} in {Reinforcement} {Learning}},
+volume = {2019-Augus},
+isbn = {978-1-72811-884-0},
+url = {https://ieeexplore.ieee.org/document/8848048/},
+doi = {10.1109/CIG.2019.8848048},
+booktitle = {2019 {IEEE} {Conference} on {Games} ({CoG})},
+publisher = {IEEE},
+author = {Harries, Luke and Lee, Sebastian and Rzepecki, Jaroslaw and Hofmann, Katja and Devlin, Sam},
+month = aug,
+year = {2019},
+keywords = {Artificial intelligence, Machine learning, Machine learning algorithms},
+pages = {1--4},
+}
+
+@article{Lehman2008,
+title = {Exploiting {Open}-{Endedness} to {Solve} {Problems} {Through} the {Search} for {Novelty}},
+number = {Alife Xi},
+journal = {Artificial Life XI},
+author = {Lehman, Joel and Stanley, Kenneth O},
+year = {2008},
+pages = {329--336},
+}
+
+@article{Miconi2018,
+title = {Differentiable plasticity: {Training} plastic neural networks with backpropagation},
+volume = {8},
+url = {https://proceedings.mlr.press/v80/miconi18a.html},
+journal = {35th International Conference on Machine Learning, ICML 2018},
+author = {Miconi, Thomas and Clune, Jeff and Stanley, Kenneth O.},
+year = {2018},
+arxiv = {1804.02464},
+isbn = {9781510867963},
+pages = {5728--5739},
 }
 
 @book{Sutton2018,
