### 2024-09-12 #7208 @kgd-al
Thanks @stheid for taking the time to review the library!
The recent commit https://github.com/kgd-al/amaze/commit/50375c662d92028eb964ff8fa2378395f1e3fbc8 to the joss branch aims to address both points.
Explicit references to maze environments have been added in the table and text and the contribution section has been expanded (https://amaze.readthedocs.io/en/joss/#contributing)

Do reach out if something else could be improved!
