+++ src/amaze/simu/simulation.py
@@ -1,3 +1,4 @@
+import pprint
 from logging import getLogger
 from pathlib import Path
 from types import SimpleNamespace
