+++ pyproject.toml
@@ -44,7 +44,7 @@ build-backend = "setuptools.build_meta"
 
 [project.optional-dependencies]
 sb3 = [
-  "stable_baselines3[extra]",
+  "stable_baselines3",
   "tensorboard",
   "graphviz"
 ]
