### 2024-10-01 7500ffa
Tweaked trajectory rendering.
