+++ src/amaze/simu/simulation.py
@@ -21,6 +21,11 @@ logger = getLogger(__name__)
 
 REWARDS = {
     "optimal": lambda length, dt: length,
+    "minimal": lambda length, dt, deadline, rewards: (
+        deadline * length
+        * (rewards.timestep
+           + min(rewards.backward, rewards.collision))
+        ) / dt,
     "compute": lambda length, dt: SimpleNamespace(
         timestep=-dt,
         backward=-1 / 10,
@@ -73,9 +78,10 @@ class Simulation:
         self.dt = 1 if self.data.outputs is OutputType.DISCRETE else 0.1
 
         sl = len(self.maze.solution)
-        self.deadline = 4 * sl / self.dt
+        self.deadline = self.deadline_factor() * sl / self.dt
         self.rewards = REWARDS["compute"](sl, self.dt)
         self.optimal_reward = REWARDS["optimal"](sl, self.dt)
+        self.minimal_reward = REWARDS["minimal"](sl, self.dt, self.deadline_factor(), self.rewards)
         self.stats = SimpleNamespace(steps=0, collisions=0, backsteps=0)
 
         self.observations = self._observations(self.data.inputs, self.data.vision)
@@ -99,6 +105,9 @@ class Simulation:
     def time(self):
         return self.timestep * self.dt
 
+    @staticmethod
+    def deadline_factor(): return 4
+
     def success(self):
         """Return whether the agent has reached the target"""
         return self.robot.cell() == self.maze.end
@@ -108,13 +117,21 @@ class Simulation:
         return self.timestep >= self.deadline
 
     def done(self):
+        """Return whether the simulation has run its course
+
+        @see :meth:`success()` and :meth:`failure()` to check the outcome
+        """
         return self.success() or self.failure()
 
     def cumulative_reward(self):
         return self.robot.reward
 
     def normalized_reward(self):
-        """Return the agent's cumulative reward in :math:`(-\\inf, 1]`"""
+        """Return the agent's cumulative reward in :math:`[-1, 1]`"""
+        r_max, r_min = self.optimal_reward, self.minimal_reward
+        r = round(2 * (self.robot.reward - r_min) / (r_max - r_min) - 1, 3)
+        assert -1 <= r <= 1, r
+        return r
         return (
             2 * int(self.success())
             - self.dt * self.stats.steps / (len(self.maze.solution) - 1)
@@ -310,7 +327,7 @@ class Simulation:
         if self.done():
             reward += self.rewards.finish
 
-        if prev_prev_cell == self.robot.cell():
+        if prev_prev_cell != self.robot.prev_cell and prev_prev_cell == self.robot.cell():
             reward += self.rewards.backward
             self.stats.backsteps += 1
 
