### 2024-10-01 e9f8ef1
Tweaked trajectory rendering.
