+++ src/amaze/visu/widgets/_trajectory_plotter.py
@@ -209,7 +209,6 @@ def _plot_trajectory_value(
 ) -> Optional[QImage]:
 
     print("[kgd-debug] ==")
-    print(trajectory)
     verbose = 1
     print("[kgd-debug] --")
 
@@ -227,83 +226,77 @@ def _plot_trajectory_value(
     if side == 0:
         side = -1 if maze.start[0] == 0 else 1
 
-    # First look for cycles
-    data = pd.DataFrame(columns=["i", "j", "cr", "steps", "value"])
     c_reward = 0
-    values = []
     solution_index = 0
     backtrack_stack = []
-    state_action = set()
-    cycle = None
+    state_action = dict()
+
+    r_min = simulation.minimal_reward
+    r_range = simulation.optimal_reward - r_min
+    def v_norm(_v): return round((v - r_min) / r_range, 3)
+
     for i, px, py, ax, ay, r in trajectory.itertuples():
         if verbose > 2:  # pragma: no cover
             print("===")
 
-        if config["cycles"]:
-            # Detect cycles
-            k = ((px, py), (ax, ay))
-            if k in state_action:
-                cycle = i
-                if verbose > 1:
-                    print("Cycle detected")
-                break
-            state_action.add(k)
-
         # Compute value
         c_reward += r
-        v = c_reward
-        if r < 0:
-            v += rewards.finish
-
-        steps = 0
+        required_steps = 0
 
         cell = (int(px), int(py))
-        next_cell = (int(px + ax), int(py + ay))
-        if verbose > 2:  # pragma: no cover
-            print(f"{cell=}, {next_cell=}")
-        if next_cell == maze.solution[solution_index + 1]:
+        if i < len(trajectory) - 1:
+            next_cell = tuple(int(_p) for _p in trajectory[["px", "py"]].iloc[i+1])
             if verbose > 2:  # pragma: no cover
-                print("Still on track")
-            solution_index += 1
-        elif next_cell == maze.solution[solution_index]:
-            if verbose > 2:  # pragma: no cover
-                print("Backward")
-            solution_index -= 1
-        else:
-            if verbose > 2:  # pragma: no cover
-                print("Off track")
-            steps += len(backtrack_stack)
-            if backtrack_stack and next_cell == backtrack_stack[-1]:
-                backtrack_stack.pop()
-            # elif cell != maze.solution[solution_index]:
+                print(f"{cell=}, {next_cell=}")
+            if cell == next_cell:
+                if verbose > 2:
+                    print("Hitting a wall")
+                pass
+            elif next_cell == maze.solution[solution_index + 1]:
+                if verbose > 2:  # pragma: no cover
+                    print("Still on track")
+                solution_index += 1
+            elif next_cell == maze.solution[solution_index]:
+                if verbose > 2:  # pragma: no cover
+                    print("Backward")
+                solution_index -= 1
             else:
-                backtrack_stack.append(cell)
-        if verbose > 2:  # pragma: no cover
-            print(len(backtrack_stack), backtrack_stack)
-        steps += len(maze.solution) - solution_index - 1
+                if verbose > 2:  # pragma: no cover
+                    print("Off track")
+                required_steps += len(backtrack_stack)
+                if backtrack_stack and next_cell == backtrack_stack[-1]:
+                    backtrack_stack.pop()
+                # elif cell != maze.solution[solution_index]:
+                else:
+                    backtrack_stack.append(cell)
[…]
