+++ src/amaze/visu/viewer.py
@@ -93,7 +93,6 @@ class MainWindow(QWidget):
 
     def __init__(self, args: Optional = None, runnable=True):
         super().__init__()
-        self.setMaximumWidth(500)
         self.args = args
 
         # holder = QWidget()
@@ -835,7 +834,7 @@ class MainWindow(QWidget):
 
     def _connect_signals(self):
         reset_maze = functools.partial(self.reset, MainWindow.Reset.MAZE)
-        for k in ["width", "height", "seed", "vision", "p_lure", "p_trap"]:
+        for k in ["width", "height", "seed", "p_lure", "p_trap"]:
             self.config[k].valueChanged.connect(reset_maze)
         for k in ["traps", "lures", "clues"]:
             c = self.config[k]
@@ -848,6 +847,7 @@ class MainWindow(QWidget):
         reset_robot = functools.partial(
             self.reset, MainWindow.Reset.ROBOT | MainWindow.Reset.CONTROL
         )
+        self.config["vision"].valueChanged.connect(reset_robot)
         for k in ["inputs", "outputs"]:
             self.config[k].currentTextChanged.connect(reset_robot)
 
@@ -911,21 +911,25 @@ class MainWindow(QWidget):
 
         viewer_options = {}
 
-        if args.robot is not None:
-            args_robot = Robot.BuildData.from_string(
-                args.robot,
-                Robot.BuildData.from_argparse(args, set_defaults=False),
-            )
-        else:
-            args_robot = Robot.BuildData.from_argparse(args, set_defaults=True)
+        defaults = not args.restore_config
 
-        if args.maze is not None:
-            args_maze = Maze.BuildData.from_string(
-                args.maze,
-                Maze.BuildData.from_argparse(args, set_defaults=False),
+        RBD = Robot.BuildData
+        if args.robot is not None:
+            args_robot = RBD.from_string(
+                args.robot,
+                RBD.from_argparse(args, set_defaults=False),
             )
         else:
-            args_maze = Maze.BuildData.from_argparse(args, set_defaults=True)
+            args_robot = RBD.from_argparse(args, set_defaults=defaults)
+
+        MBD = Maze.BuildData
+        if args.maze is not None:
+            args_maze = MBD.from_string(
+                args.maze,
+                MBD.from_argparse(args, set_defaults=False),
+            )
+        else:
+            args_maze = MBD.from_argparse(args, set_defaults=defaults)
 
         if args.restore_config:
             for k in MazeWidget.default_config().keys():
@@ -934,11 +938,13 @@ class MainWindow(QWidget):
                 viewer_options[k] = b
 
             self._init_from_robot_build_data(
-                Robot.BuildData(**_try("robot", default={})).override_with(args_robot)
+                RBD.from_string(_try("robot", default={}))
+                   .override_with(args_robot)
             )
 
             self._init_from_maze_build_data(
-                Maze.BuildData(**_try("maze", default={})).override_with(args_maze)
+                MBD.from_string(_try("maze", default={}))
+                   .override_with(args_maze)
             )
 
             config.beginGroup("sections")
@@ -964,8 +970,8 @@ class MainWindow(QWidget):
         config.setValue("pos", self.pos())
         config.setValue("size", self.size())
 
-        config.setValue("maze", self.maze_data().__dict__)
-        config.setValue("robot", self._robot_data().__dict__)
+        config.setValue("maze", self.maze_data().to_string())
+        config.setValue("robot", self._robot_data().to_string())
 
         config.beginGroup("show")
         for k in MazeWidget.default_config().keys():
