+++ src/amaze/visu/viewer.py
@@ -886,79 +886,86 @@ class MainWindow(QWidget):
         return QSettings("kgd", "amaze")
 
     def _restore_settings(self, args):
-        if args.restore_config:
-            config = self._settings()
-            logger.info(f"Loading configuration from {config.fileName()}")
+        try:
+            if args.restore_config:
+                config = self._settings()
+                logger.info(f"Loading configuration from {config.fileName()}")
 
-        def _try(name, f_=None, default=None):
-            if (v_ := config.value(name.replace("_", "/"))) is not None:
-                return f_(v_) if f_ else v_
-            else:
-                return default
+            def _try(name, f_=None, default=None):
+                if (v_ := config.value(name.replace("_", "/"))) is not None:
+                    return f_(v_) if f_ else v_
+                else:
+                    return default
 
-        if args.restore_config:
-            _try("pos", lambda p: self.move(p))
-            _try("size", lambda s: self.resize(s))
+            if args.restore_config:
+                _try("pos", lambda p: self.move(p))
+                _try("size", lambda s: self.resize(s))
 
-        if (w := args.width) is not None or (w := self.width()) < 500:
-            w_, h_ = self.width(), self.height()
-            self.resize(w, w * h_ // w_)
+            if (w := args.width) is not None or (w := self.width()) < 500:
+                w_, h_ = self.width(), self.height()
+                self.resize(w, w * h_ // w_)
 
-        def restore_bool(s, k__):
-            b_ = bool(int(s))
-            self.config[k__].setChecked(b_)
-            return b_
+            def restore_bool(s, k__):
+                b_ = bool(int(s))
+                self.config[k__].setChecked(b_)
+                return b_
 
-        viewer_options = {}
+            viewer_options = {}
 
-        defaults = not args.restore_config
+            defaults = not args.restore_config
 
-        RBD = Robot.BuildData
-        if args.robot is not None:
-            args_robot = RBD.from_string(
-                args.robot,
-                RBD.from_argparse(args, set_defaults=False),
-            )
-        else:
-            args_robot = RBD.from_argparse(args, set_defaults=defaults)
-
-        MBD = Maze.BuildData
-        if args.maze is not None:
-            args_maze = MBD.from_string(
-                args.maze,
-                MBD.from_argparse(args, set_defaults=False),
-            )
-        else:
-            args_maze = MBD.from_argparse(args, set_defaults=defaults)
-
-        if args.restore_config:
-            for k in MazeWidget.default_config().keys():
-                k_ = "show_" + k
-                b = _try(k_, functools.partial(restore_bool, k__=k_))
-                viewer_options[k] = b
-
-            self._init_from_robot_build_data(
-                RBD.from_string(_try("robot", default={}))
-                   .override_with(args_robot)
-            )
-
-            self._init_from_maze_build_data(
-                MBD.from_string(_try("maze", default={}))
-                   .override_with(args_maze)
-            )
-
-            config.beginGroup("sections")
-            for k in self.sections:
-                self.sections[k].set_collapsed(
-                    bool(int(config.value(k.value.lower(), False)))
+            RBD = Robot.BuildData
+            if args.robot is not None:
+                args_robot = RBD.from_string(
+                    args.robot,
+                    RBD.from_argparse(args, set_defaults=False),
                 )
-            config.endGroup()
+            else:
+                args_robot = RBD.from_argparse(args, set_defaults=defaults)
 
-        else:
-            self._init_from_robot_build_data(args_robot)
-            self._init_from_maze_build_data(args_maze)
+            MBD = Maze.BuildData
+            if args.maze is not None:
+                args_maze = MBD.from_string(
+                    args.maze,
+                    MBD.from_argparse(args, set_defaults=False),
+                )
+            else:
[…]
