### 2024-11-29 #7208 @ymzayek
Hi @auroreloisy and @inpefess,

Would you be willing to assist in reviewing this submission for JOSS (Journal of Open Source Software)?

AMaze: a benchmark generator for sighted maze-navigating agents

You can find more information at the top of this Github issue. If you have any questions please let me know.
Thank you in advance!
