### 2024-12-22 kgd-al/amaze#3 @inpefess
JOSS paper review
Thank you for drafting the paper https://github.com/openjournals/joss-reviews/issues/7208#issuecomment-2348291626

Here are my suggestions and remarks for the text.

The content of the Summary and Statement of Need sections should probably be swapped or edited. According to the JOSS policy, the Summary should be "a clear description of the high-level functionality and purpose of the software for a diverse, non-specialist audience." In the draft, the Statement of Need explains what AMaze is, and the Summary talks about the need. It should be vice versa.

Figure 1. The description reads "Every maze can be converted to and from a human-readable string". I don't think there is a way to convert a maze to such a string, because the string contains probabilities which can only be estimated based on one particular maze. It would probably be better to write something like "AMaze generates mazes based on human-readable strings".

Figure 4. A better explanation is needed. What is N? What is Time exactly? What is Image as an input type and how is it different from Continuous? Which Control is considered Extensive? What is Mode and Lua? Why is Toy Text marked as providing no control over the environment when one can set a custom map for a FrozenLake game?

line 11: "By using visual cues" - should probably be clues

line 20: "Every maze can be described by human-readable string as illustrated in Figure 1" - since there is no definition of a maze in the paper, I would discourage any theoretical claim about "every maze". If I got it right, the author probably wanted to say something like "Amaze generates mazes based on human-readable strings..."

lines 36-37: "the library provides, for any given maze M, two dedicated metrics, the surprisingness S_M and deceptiveness D_M" - I could not find package documentation on how to compute these metrics for a given maze. Please add the documentation or drop this claim.

line 46: "a single bit of memory (the direction" - the phrase might be somewhat misleading (1 bit is not enough to store four directions), please reword

line 62: "Computational efficiency" - AMaze is a maze generator (or benchmark generator). Does computational efficiency mean it can generate many mazes in a short time? If so, why is this important? Please elaborate.

line 63: "intuitive understanding of an agent’s behavior" - I'm not sure I understand the meaning of this phrase, and I don't find any support for this claim in the following paragraphs of the draft.

line 71: "with different vision sizes (11, 15, 21)" - please include an explanation of what is "vision size" before or after this line

lines 73-74: "to give sufficient data for a fair comparison" - I suggest supporting this claim with some statistical criteria calculations. I'm not claiming the contrary, but for me, it's not immediately obvious that the amount of data was indeed sufficient.
