### 2024-12-22 kgd-al/amaze#4 @inpefess
JOSS paper: Reproducibility
Thank you for creating AMaze, it's really amazing :smiley:

According to JOSS policy, the paper should satisfy the criterion of "Reproducibility: If the paper contains original results, results are entirely reproducible by reviewers." https://github.com/openjournals/joss-reviews/issues/7208

There is indeed an original result: an experimental comparison of AMaze with similar software. I suggest adding a documentation page about reproducing these experiments.

Another result I think should be reproducible is collecting data for Figure 2 (500000 mazes of different classes with their metrics).
