### 2024-12-22 kgd-al/amaze#5 @inpefess
JOSS paper: list of authors
Kudos to all contributors to AMaze!

As a reviewer of your JOSS paper draft https://github.com/openjournals/joss-reviews/issues/7208, I should answer whether "the full list of paper authors seems appropriate". According to the git commit history, there is only one code author, but the paper lists three authors. I believe adding a file explaining the roles of non-code contributors might help. What do you think?
