### 2024-12-22 #7208 @inpefess
@kgd-al Thank you and your co-authors for open-sourcing AMaze! I've done the first round of my review, so please find the corresponding issues linked to this one.
@ymzayek Since I asked to add scripts for reproducing experiments and some text edits, I will need one more round to run the code and read the draft after the author addresses my remarks.
