### 2025-01-06 kgd-al/amaze#5 @kgd-al
Both other contributors have been involved in an administrative and management capacity.
Though I am the sole developer, I would find it rude not to acknowledge their participation.
