### 2025-01-06 kgd-al/amaze#3 @kgd-al
Thank you for taking the time to do such a thorough read.
Here is my take on it, in the same order:
- I agree about the statement of need. I have split it between the statement itself and a feature description.
- Every maze can be converted: I highlighted that it is done through the API. Manual conversion from an image is possible but painfully long.
- I added more text alongside Figure 4 to make it clearer. However for Toy Text, only Frozen Lake allows for manual customization (added in the text but not in the graph).

-- to be continued --
