+++ joss/paper.md
@@ -26,8 +26,6 @@ date: 21 June 2024
 bibliography: paper.bib
 ---
 
-<!-- Add lab2D and Maze explorer to the table. And some text to go along with it. Mention Lehman2011 and more (for benchmark-less mazes) -->
-
 # Summary
 
 The need to provide fair comparisons between agents, especially in the field of Reinforcement Learning, has led to a plethora of benchmarks.
@@ -39,14 +37,20 @@ By using visual cues in a maze-navigation task, the library empowers researchers
 
 AMaze is a pure-Python package with an emphasis on the easy and intuitive generation, evaluation and analysis of mazes.
 Its primary goal is to provide a way to quickly generate mazes of targeted difficulty, e.g., to test a Reinforcement Learning algorithm.
+By modeling loosely embodied robots with three distinct input/output spaces, AMaze makes it possible to prototype agent-centric scenarios of decision making, pattern recognition and general behavior through exposition to a wide array of contexts.
+
+[^1]: [https://amaze.readthedocs.io/en/latest/](https://amaze.readthedocs.io/en/latest/)
+
+![A sample maze from the AMaze library. In the API, every maze can be converted to and from a human-readable string where each underscore-separated component describes one of its facets. The *seed* seeds the random number generator used for the paths and stochastic placement of *lures* and *traps*. These have a specific probability, shape and/or value and may be specified multiple times to increase the complexity, as described in the documentation[^1]\label{fig:maze}](../docs/latex/maze/light-wide.png)
+
+# Features
+
 Users of AMaze have two main components to take into consideration: mazes and agents.
 These are introduced below with more details available in the documentation[^1].
 
-![A sample maze from the AMaze library. Every maze can be converted to and from a human-readable string where each underscore-separated component describes one of its facets. The *seed* seeds the random number generator used for the paths and stochastic placement of *lures* and *traps*. These have a specific probability, shape and/or value and may be specified multiple times to increase the complexity.\label{fig:maze}](../docs/latex/maze/light-wide.png)
-
 ## Mazes
 
-Every maze can be described by human-readable string as illustrated in \autoref{fig:maze}, where every component is optional with sensible default values.
+Every maze can be described by human-readable string as illustrated in \autoref{fig:maze}, where every component is optional.
 The *seed* is used in the random number generator responsible for: a) the depth-first search that creates the paths and b) the stochastic placement of the *lures* and *traps*.
 As will be detailed below, agents only see a single cell at a time making intersections impossible to handle without additional information.
 *Clues* provide such an information by helpfully pointing towards the correct direction.
@@ -69,9 +73,7 @@ which, informally, account for the likelihood of encountering different states (
 Through these metrics, experimenters can make an informed decision about the level of complexity of the mazes they use.
 As illustrated by the distributions of $S_M$ and $D_M$, sampled from 500'000 mazes across all five classes (\autoref{fig:complexity}), the space of all possible mazes is both diverse and arbitrarily complex.
 
-![Distribution of Surprisingness $S_M$ versus Deceptiveness $D_M$ across 500'000 unique mazes from all five different classes. Outlier mazes are depicted in the borders to illustrate the underlying Surprisingness (right column) or lack thereof (left column).\label{fig:complexity}](../docs/latex/complexity/light.png)
-
-[^1]: [https://amaze.readthedocs.io/en/latest/](https://amaze.readthedocs.io/en/latest/)
[…]
