### 2025-01-07 53fda4a
More progress on issue #3 and some fixes
