+++ joss/make.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
 cd joss/inara
-make pdf ARTICLE=../paper.md || exit 1
+make pdf ARTICLE=../paper.md TARGET_FOLDER=.. || exit 1
 
-pgrep -f paper.pdf >/dev/null || xdg-open publishing-artifacts/paper.pdf
+pgrep -f paper.pdf >/dev/null || xdg-open ../paper.pdf
