+++ src/amaze/misc/resources.py
@@ -527,18 +527,18 @@ def __generator__alien(lightness: float, size: int):
     img, painter = _image(size, lightness)
 
     path = QPainterPath()
-    path.moveTo(0.0, -.5)
-    path.lineTo(0.4, -.5)
+    path.moveTo(0.0, -0.5)
+    path.lineTo(0.4, -0.5)
     path.lineTo(1.0, 0.0)
-    path.lineTo(0.4, +.5)
-    path.lineTo(0.0, +.5)
-    path.lineTo(0.0, +.125)
-    path.lineTo(0.4, +.25)
+    path.lineTo(0.4, +0.5)
+    path.lineTo(0.0, +0.5)
+    path.lineTo(0.0, +0.125)
+    path.lineTo(0.4, +0.25)
     path.lineTo(0.6, 0.0)
-    path.lineTo(0.4, -.25)
-    path.lineTo(0.0, -.125)
+    path.lineTo(0.4, -0.25)
+    path.lineTo(0.0, -0.125)
     path.closeSubpath()
-    painter.translate(0, .5)
+    painter.translate(0, 0.5)
     painter.fillPath(path, __pen_color(lightness))
 
     painter.end()
@@ -577,18 +577,18 @@ def __generator__star(lightness: float, size: int):
     painter.drawEllipse(QRectF(-0.25, -0.25, 0.5, 0.5))
 
     path = QPainterPath()
-    path.moveTo(0, .05)
-    path.lineTo(.35, .05)
-    path.lineTo(.35, .1)
-    path.lineTo(.5, .0)
-    path.lineTo(.35, -.1)
-    path.lineTo(.35, -.05)
-    path.lineTo(0, -.05)
+    path.moveTo(0, 0.05)
+    path.lineTo(0.35, 0.05)
+    path.lineTo(0.35, 0.1)
+    path.lineTo(0.5, 0.0)
+    path.lineTo(0.35, -0.1)
+    path.lineTo(0.35, -0.05)
+    path.lineTo(0, -0.05)
     path.closeSubpath()
 
     for i in range(8):
         painter.save()
-        painter.rotate(i*45)
+        painter.rotate(i * 45)
         painter.fillPath(path, __pen_color(lightness))
         painter.restore()
 
