### 2025-01-07 kgd-al/amaze#3 @kgd-al
-- continuing --
 - line 11. That is intended. Cues are placed in the maze to prompt action. Some of them are actual clues (i.e. helpful).
 - line 20. I meant any maze generated by the library.
 - line 36. They were meant to be accessible through the maze stats but this was neither implemented nor documented. Thank you for catching it! There is now a dedicated statement in the [basics](https://amaze.readthedocs.io/en/joss/usage/basics.html) tutorial. Also the API now makes explicit mention of it (https://amaze.readthedocs.io/en/joss/_autosummary/amaze.simu.simulation.html#amaze.simu.simulation.Simulation.compute_metrics).
 - line 46. Agreed. Reformulated.
 - line 62. Elaborated.
 - line 63. Made it clearer (hopefully).
 - line 71. Reformulated as input image size
 - line 73. Given the wide variability of sub-environment counts, I would welcome ideas on a viable criteria.

@inpefess I hope this solves/answers most of your suggestions and remarks.
Do tell me if you find anything still amiss and thanks again for taking the time!
