+++ docs/src/misc.rst
@@ -2,10 +2,16 @@
 Miscellaneous
 =============
 
-Todo list
-^^^^^^^^^
+Benchmark comparison
+^^^^^^^^^^^^^^^^^^^^
 
-.. todolist::
+In the companion [JOSS](https://joss.theoj.org/)
+[paper](https://joss.theoj.org/papers/bc6892f6873808ced10bc8c4c222e635)
+AMaze is compared against a number of other benchmarks from the Python ecosystem.
+The scripts (Python and Shell) to reproduce a under docs/latex/benchmarking:
+
+- `generate_all.sh` iterates over every declared package and proceeds, if necessary, to
+  installing it before 
 
 Build Errors
 ^^^^^^^^^^^^
