+++ docs/src/development.rst
@@ -1,0 +1,12 @@
+===================
+Development details
+===================
+
+Todo list
+^^^^^^^^^
+
+.. todolist::
+
+Build Errors
+^^^^^^^^^^^^
+.. include:: _autogen/errors.rst
