### 2025-01-07 2b7d659
Reproducibility of benchmark data (Issue #4)
