### 2025-01-08 kgd-al/amaze#4 @kgd-al
Reproducibility of the comparison with similar software has been [documented](https://amaze.readthedocs.io/en/joss/misc.html#benchmark-comparison).

The script for the .5M distribution of complexity is nowhere to be found so recreating will take some time.
