+++ docs/latex/complexity/generate.py
@@ -1,12 +1,12 @@
 #!/bin/env python3
-import itertools
-import math
+
 import pprint
 import signal
 from pathlib import Path
+from random import Random
 
 import pandas as pd
-from rich.progress import track
+from rich.progress import Progress
 
 from amaze import Maze, Sign, InputType, Simulation
 from amaze.simu import MazeMetrics
@@ -15,20 +15,35 @@ from amaze.simu.types import MazeClass
 persistent_data = Path(__file__).parent.joinpath("data.csv")
 
 try:
+    # raise FileNotFoundError()
     df = pd.read_csv(persistent_data, index_col=0)
 except FileNotFoundError:
-    df = pd.DataFrame(columns=["ID", "Class", "Deceptiveness", "Surprisingness"])
+    df = pd.DataFrame(columns=["Class", "Deceptiveness", "Surprisingness"])
+    df.index.name = "ID"
 
-seeds = range(200)
-sizes = [5, 10, 20, 50]
-signs = [Sign(value=v) for v in [.25, .5, .75, 1]]
-clues = [[], [signs[3]], [signs[1], signs[3]], signs]
-lures = [[], [signs[0]], [signs[0], signs[2]], signs]
-traps = [[], [signs[1]], [signs[0], signs[2]], signs]
-probas = [0, .25, .5, .75, 1]
+n = 100000
 
-items = [seeds, clues, lures, traps, probas, probas, sizes]
-total = math.prod([len(i) for i in items])
+
+def build_data(_rng: Random, _m_class: MazeClass):
+    def _signs(): return [Sign(value=.5 * _rng.random() + .5) for _ in range(_rng.randint(2, 4))]
+    _size = _rng.randint(2, 50)
+    _clues = _signs() if _m_class.value & 0b001 else []
+    _p_lure, _lures = 0, []
+    _p_trap, _traps = 0, []
+    if (_m_class.value & 0b010) or _m_class.value is MazeClass.TRIVIAL:
+        _p_lure = .75 * _rng.random() + .25
+        _lures = _signs()
+    if _m_class.value & 0b100:
+        _p_trap = .75 * _rng.random() + .25
+        _traps = _signs()
+    return Maze.BuildData(
+        seed=_rng.randint(0, 2 ** 32),
+        width=_size, height=_size,
+        clue=_clues,
+        p_lure=_p_lure, lure=_lures,
+        p_trap=_p_trap, trap=_traps,
+        unicursive=_m_class is MazeClass.TRIVIAL,
+    )
 
 
 def handler(*_):
@@ -40,32 +55,56 @@ def handler(*_):
 for s in [signal.SIGINT, signal.SIGTERM]:
     signal.signal(signal.SIGINT, handler)
 
-
+nmc = len(MazeClass) - 1
+total = nmc * n
 if len(df) < total:
-    for seed, clue, lure, trap, p_lure, p_trap, size in track(
-            itertools.product(*items), description="Processing", total=total):
+    queried, generated, skipped, kept = 0, 0, 0, 0
+    sizes = {c: 0 for c in MazeClass if c is not MazeClass.INVALID}
+    m_classes = list(sizes.keys())
 
-        bd = Maze.BuildData(
-            seed=seed,
-            width=size, height=size,
-            clue=clue,
-            p_lure=p_lure, lure=lure,
-            p_trap=p_trap, trap=trap,
-            unicursive=(len(clue) == 0)
-        )
-        name = bd.to_string()
-        if name not in df.index:
-            maze = Maze.generate(bd)
-            m_class = maze.maze_class()
-            assert m_class is not MazeClass.INVALID, \
[…]
