### 2025-01-11 kgd-al/amaze#5 @inpefess
Sure, that's why I proposed to acknowledge their participation. Although acknowledging non-code contributions in code repos is a widespread practice, there is no standard way. You can add a section to the ReadMe file or create a separate file, for example.
