### 2025-01-11 kgd-al/amaze#3 @inpefess
Great! Where can I have a look at the updated paper draft?
