### 2025-01-11 kgd-al/amaze#4 @inpefess
I tried to run the script (fresh clone, commit a702f967208e7ccb19ac56975e1591caf549285e)

```sh
cd ./docs/latex/benchmarking
./generate_all.sh
```

but got an error: `./generate_all.sh: line 35: .../amaze/docs/latex/benchmarking/___cache___//log: No such file or directory`

For the second script - no problem, take your time.
