### 2025-01-13 kgd-al/amaze#5 @kgd-al
I better understand your point now. I added a statement in the acknowledgement section of the README.
