### 2025-01-13 kgd-al/amaze#3 @kgd-al
Easiest solution is to use the command `@editorialbot generate pdf` in the review issue tracker (https://github.com/openjournals/joss-reviews/issues/7208)
