+++ docs/latex/complexity/format.py
@@ -1,43 +1,133 @@
 import math
+import sys
 from pathlib import Path
 from random import Random
 
+from PIL import Image
+import matplotlib.pyplot as plt
+import matplotlib.colors as mcolors
+from matplotlib.offsetbox import OffsetImage, AnnotationBbox
+from scipy import spatial
+
 import pandas as pd
 
-import matplotlib.pyplot as plt
+import matplotlib.image as image
 import seaborn as sns
 
+from amaze import amaze_main
 from amaze.simu.types import MazeClass
 
-df_file = Path(__file__).parent.joinpath('data.csv')
+dpi = 300
+
+root = Path(__file__).parent
+df_file = root.joinpath('data.csv')
 df = pd.read_csv(df_file, index_col=0)
 
 print(df)
 hue_order = [c.name.capitalize() for c in list(reversed(MazeClass))[1:]]
 
-if False:
-    kwargs = dict(x="Surprisingness", y="Deceptiveness", hue="Class", hue_order=hue_order)
+if len(sys.argv) > 1:
+    print("Fast-mode: sampling 10000 points from the data")
+    df = df.sample(10000, random_state=1)
 
-    sns.set_style("darkgrid")
-    g = sns.jointplot(data=df, **kwargs,
-                      kind="kde", cut=0, levels=10, thresh=.01, bw_adjust=1,
-                      fill=True, alpha=.5,
-                      marginal_kws=dict(
-                          multiple="stack", alpha=.25, linewidth=.1,
-                          bw_adjust=1,
-                          warn_singular=False
-                      ),
-                      warn_singular=False)
-    g.ax_joint.legend_.set_title("Maze class")
-    g.ax_joint.autoscale(enable=True, tight=True)
+# Sample points for readable scatterplot
+# For Trivial and Simple: get the min-max along S (D always equals 0)
+# For Lures, Traps and Complex: get the convex hull
+# Then complement as needed
+sample_size = 1000
+sample_df = df[df.Class == "NOTHING"]
+for m_class in hue_order:
+    c_df = df[df.Class == m_class]
+    if len(c_df["Deceptiveness"].unique()) == 1:
+        hull_idx = c_df.Surprisingness.agg(['idxmin', 'idxmax'])
+    else:
+        hull_idx = spatial.ConvexHull(c_df[["Surprisingness", "Deceptiveness"]]).vertices
+        hull_idx = c_df.index[hull_idx]
+    hull_df = c_df.loc[hull_idx]
+    c_df = c_df.drop(hull_idx)
+    sample_df = pd.concat([
+        sample_df, hull_df,
+        c_df.sample(sample_size // 5 - len(hull_df), random_state=0)
+    ])
 
-    sns.scatterplot(data=df.sample(n=100, random_state=0), **kwargs, size=.1,
-                    ax=g.ax_joint, legend=False)
+print("Sample:", sample_df)
 
-    plt.savefig('foo.pdf')
+# Plot kde of data + extracted sample
+kwargs = dict(x="Surprisingness", y="Deceptiveness", hue="Class", hue_order=hue_order)
+sns.set_style("darkgrid")
+g = sns.jointplot(data=df, **kwargs,
+                  kind="kde", cut=0, levels=10, thresh=.01, bw_adjust=.7,
+                  fill=True, alpha=.5,
+                  marginal_kws=dict(
+                      multiple="stack", alpha=.25, linewidth=.1,
+                      bw_adjust=.8,
+                      warn_singular=False
+                  ),
+                  warn_singular=False)
+g.ax_joint.legend_.set_title("Maze class")
+g.ax_joint.autoscale(enable=True, tight=True)
 
-if True:
-    def eqd(a, b): return math.sqrt(a*a+b*b)
[…]
