### 2025-01-13 2462db7
Fixed missing mkdirs in benchmarking/generate
