### 2025-01-13 kgd-al/amaze#4 @kgd-al
Thanks for trying it out. This particular mistake should be fixed. For your next attempt, could you use the latest commit [2462db7](https://github.com/kgd-al/amaze/commit/2462db7b48206e4c4166b8a321939d89e0875f3f) from the joss branch?
