### 2025-01-14 d6440e6
Added script for complexity distribution (Issue #4)
