+++ docs/latex/compile.sh
@@ -31,9 +31,9 @@ compile_ld(){
 }
 
 compile_ld maze
+compile maze light-wide
 compile_ld agents
 compile_ld complexity
 
-compile maze light-wide
 
 montage -geometry +100+0 agents/light-{1,3}.png agents/light-1-3.png
