### 2025-01-14 kgd-al/amaze#4 @kgd-al
@inpefess The script to generate the distribution plot has been added in the latest commit.
