### 2025-01-15 cc0bd06
Fixed --cell-width input ignored
