+++ src/amaze/visu/widgets/maze.py
@@ -210,7 +210,15 @@ class MazeWidget(QLabel):
         super().update()
 
     @staticmethod
-    def __compute_size(maze, width, square: bool = False):
+    def __compute_size(maze: Maze,
+                       width: Optional[int] = None,
+                       cell_width: Optional[int] = None,
+                       square: bool = False):
+        cell_width = cell_width or 15
+        c_width = maze.width * cell_width
+        width = width or c_width
+        width = max(width, c_width)
+
         if square:
             return width, width
         else:
@@ -238,28 +246,57 @@ class MazeWidget(QLabel):
 
         painter.end()
 
-    def render_to_file(self, path, width: Optional[int] = None):
-        """Render this widget to a file"""
-        img, painter = self._image_drawer(width=width)
-        self.render_onto(painter, width=width)
+    def render_to_file(self, path,
+                       width: Optional[int] = None,
+                       cell_width: Optional[int] = None):
+        """Render this widget to a file
+
+        :param path: destination path for the resulting image
+        :param width: requested width of the image
+        :param cell_width: requested width of a single cell
+
+        :return: Whether saving was successful
+        """
+        img, painter = self._image_drawer(width=width, cell_width=cell_width)
+        self.render_onto(painter, width=width, cell_width=cell_width)
         return self._save_image(path, img, painter)
 
-    def pretty_render(self, width: Optional[int] = None):
+    def pretty_render(self,
+                      width: Optional[int] = None,
+                      cell_width: Optional[int] = None):
+        """Render this widget to an image
+
+        :param width: requested width of the image
+        :param cell_width: requested width of a single cell
+
+        :return: The generated image
+        """
         img, painter = self._image_drawer(width=width)
         self.render_onto(painter, width=width)
         painter.end()
         return img
 
     @classmethod
-    def static_render_to_file(cls, maze: Maze, path: Path, size=None, **kwargs):
+    def static_render_to_file(cls,
+                              maze: Maze, path: Path,
+                              width: Optional[int] = None,
+                              cell_width: Optional[int] = None,
+                              **kwargs):
         """Render the provided maze to a file (in png format).
+
+        :param maze: Maze to render
+        :param path: destination path for the resulting image
+        :param width: requested width of the image
+        :param cell_width: requested width of a single cell
 
         Additional arguments refer to the rendering configuration, see
         :meth:`default_config`.
+
+        :return: Whether saving was successful
         """
-        width, height = cls.__compute_size(maze, size or maze.width * 15)
+        width, height = cls.__compute_size(maze, width, cell_width)
         scale = width / maze.width
-        config = cls.default_config()
+        config = kwargs.get("config", None) or cls.default_config()
         dark = kwargs.get("dark", True)
         config.update(
             dict(
@@ -287,7 +324,9 @@ class MazeWidget(QLabel):
 
         QtPainter(painter, config).render(maze, config)
 
-    def render_onto(self, painter: QPainter, width: Optional[int] = None):
+    def render_onto(self, painter: QPainter,
+                    width: Optional[int] = None,
+                    cell_width: Optional[int] = None,):
         if self._config["robot"]:
             robot = self._robot.to_dict()
         else:
@@ -308,8 +347,11 @@ class MazeWidget(QLabel):
             )
         )
 
-        if width is not None:
+        if cell_width is not None:
+            config["scale"] = cell_width
+        elif width is not None:
             config["scale"] = width / self._maze.width
+        print("[kgd-debug]", "render_onto", config)
 
         self.__render(painter, self._maze, self._maze.height * config["scale"], config)
 
@@ -344,15 +386,16 @@ class MazeWidget(QLabel):
 
     def _image_drawer(
         self,
[…]
