+++ docs/src/misc.rst
@@ -25,13 +25,28 @@ The scripts (Python and Shell) to reproduce a under `docs/latex/benchmarking`:
   Note that the one used for the aforementioned article is available under
   `docs/latex/benchmarking/table.csv`
 
+.. warning::
+
+    Installing all of these libraries plus their dependencies *will* take quite some time
+    and disk space.
+    It is *highly* recommended to do so in a virtual environment on a sufficiently robust
+    machine.
 
 Complexity space
 ^^^^^^^^^^^^^^^^
 
 In the README, a figure presenting the complexity of 500'000 mazes is used to showcase the
 diversity of the underlying space.
-Such a figure can be generated by????
+Such a figure can be generated by the scripts under `docs/latex/complexity/`:
+
+- `generate.py` will create 100'000 mazes of Trivial, Simple, Lures, Traps and Complex classes,
+  each.
+  Results are stored in the sibling file `data.csv`
+
+- `format.py` will generate most of the visuals (kde, image overlay)
+
+- `complexity.tex` needs to be compiled for the inset images (it is recommended to do so
+  from the main compiler at `docs/latex/compile.sh`
 
 Build Errors
 ^^^^^^^^^^^^
