+++ docs/latex/complexity/complexity.tex
@@ -4,7 +4,7 @@
 \begin{document}
 \begin{tikzpicture}[remember picture, spy using outlines={size=2 cm, connect spies, magnification=4, circle}]
  \begin{localtoimage}{distributions}{width=\textwidth}
-  \spy [red] on (-5.5,.7) in node at (-4.5, 1.8);
+  \spy [red] on (-5,.6) in node at (-4.5, 1.8);
   \spy [red] on (5.5, 2.4) in node at (4.3, 1);
   \spy [red] on (3.7,-2.3) in node at (5, -1);
  \end{localtoimage}
