### 2025-01-19 kgd-al/amaze#4 @inpefess
> The script to generate the distribution plot has been added in the latest commit.

Please specify the script's name in the documentation, it's still "Such a figure can be generated by????" now https://amaze.readthedocs.io/en/joss/misc.html#complexity-space
