### 2025-01-19 kgd-al/amaze#4 @inpefess
Now have a different error in the first script

./generate_all.sh: line 41: line: command not found
