### 2025-01-19 #7208 @inpefess
@ymzayek While checking the citations, I found [this pre-print](https://arxiv.org/abs/2411.13072) by the same authors where they write: "In this work, we introduce AMaze". This pre-print was submitted to Frontiers in Artificial Intelligence around 20 Nov 2024 while this submission's review was underway. As far as I understand, JOSS does not prohibit co-publication of a paper describing software along with a paper describing results obtained by using this software (see [here](https://joss.readthedocs.io/en/latest/submitting.html#co-publication-of-science-methods-and-software)), but suggests "that authors indicate whether related publications (published, in review, or nearing submission) exist as part of submitting to JOSS". After examining this thread, I couldn't find any indications of the other submission, so I assume you got it by other means.

After comparing this submission with the other one, I got the impression they both described AMaze (and not, e.g. some novel research findings obtained using it). While I'm sorry for learning it so late, I should ask you to clarify how this fits the JOSS co-publication policy before I can continue with the review.
