### 2025-01-20 54ff78b
Fixed documentation build error. Added short version of line
