+++ docs/src/index.rst
@@ -7,8 +7,6 @@
     usage/index
     api
 
-    complexity
-
     misc
     development
 
