### 2025-01-20 kgd-al/amaze#4 @kgd-al
Thanks for pointing it out. The documentation just was not building. It is now fixed.
For the script: yes that is one of my essential utilities, so much so that I forgot to include a standalone version. Also fixed.
