### 2025-01-20 #7208 @kgd-al
There is indeed a related article (currently) under review with Frontiers but it explores the use of AMaze for a human-in-the-loop RL training scheme. The fact that both submissions are parallel made it so that I had to explain the framework in both. Furthermore that submission happened mid October, 3 months after JOSS's review started.
@ymzayek Please confirm that all is good or how to make it so.
@inpefess Thank you for your thoroughness.
