+++ docs/latex/benchmarking/workers/levdoom_worker.py
@@ -15,6 +15,7 @@ def _evaluate(env_name):
             env.reset()
 
     env.close()
+    return True
 
 
 def process(df):
