### 2025-03-10 #7208 @kgd-al
Hello @crvernon, @ymzayek and @inpefess,
The review process has been deadlocked for almost two months because of this parallel article. Can I provide additional information to help the situation move forward? If so, please do ask!
