### 2025-03-24 #7208 @kgd-al
Hello @crvernon,
To (hopefully) help you get a quicker understanding of the situation, here is a timeline of events:
- 01/07/2024, start of the pre-review
- 10/09/2024, end of the pre-review and feedback from first reviewer
- 13/09/2024, go from first reviewer no significant progress from the second
- 15/10/2024, submission of an article to Frontiers (which uses AMaze for its experiments)
-    11/2024, search for a new second reviewer
- 29/11/2024, new second reviewer assigned
- 22/12/2024, feedback from second reviewer
- 10/01/2025, second reviewer raises worries about parallel article
- 30/01/2025, acknowledgement of situation by editor, request for clarification from managing editor in chief
- 26/02/2025, Frontiers article moves from review to accepted

Some links:
- Article on Frontiers: https://www.frontiersin.org/journals/artificial-intelligence/articles/10.3389/frai.2025.1511712/abstract
- Corresponding pre-print: https://arxiv.org/abs/2411.13072

Please do not hesitate to contact me should you need any clarification!
