### 2025-04-10 #7208 @kgd-al
Hello @crvernon,

It has now been a month since we last talked. Have you reached a conclusion regarding the validity of this article in relation to the (now published) article in Frontiers?
