### 2025-04-11 #7208 @kgd-al
@crvernon Thanks for checking.
For the Frontier paper the reviews were focused on the comparison between different training regimes and RL algorithms. I cannot state whether any of the reviewers looked at the code but no feedback was provided (nor expected) on that front.

Also, while I agree that the titles are similar, they are not identical:
- AMaze: a benchmark generator for sighted maze-navigating agents (JOSS)
- AMaze: an intuitive benchmark generator for fast prototyping of generalizable agents (Frontiers)
