### 2025-04-23 #7208 @kgd-al
@ymzayek Well back to the drawing board.
abhiramm7 or auroreloisy maybe be available now.

Out of curiosity, is this kind of flow usual for a JOSS article? It seems to be quite... eventful.
