### 2025-05-06 #7208 @ymzayek
Hello, @kgd-al sorry about the lack of activity. I can relaunch some previous requests and let me know if you have any new people in mind.

In general this review has had a stroke of bad luck with timing and events. I am also no longer an editor with JOSS but was hopeful that we could finish this even after I stopped but that was before the delay caused by the frontiers submission.

> Out of curiosity, is this kind of flow usual for a JOSS article? It seems to be quite... eventful.

It can happen very quickly/smoothly sometimes but given that people are volunteering their time it's not unheard of that a review gets stuck.
