### 2025-05-23 #7208 @ymzayek
Hello @OlgerSiebinga  and @danielskatz,

Would you be willing to assist in reviewing this submission for JOSS (Journal of Open Source Software)?

AMaze: a benchmark generator for sighted maze-navigating agents

You can find more information at the top of this Github issue and on the pre-review issue (https://github.com/openjournals/joss-reviews/issues/6945).

There is already one review finished on this submission and we need at least one more review.

Thank you in advance!
