### 2025-11-10 #7208 @crvernon
🥳 Congratulations on your new publication @kgd-al ! Many thanks to @stheid for your time, hard work, and expertise!! JOSS wouldn't be able to function nor succeed without your efforts.
